[FIX] missing OPENMS_DLLAPI, duplicate include, conversion warning
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/branches/Release1.10@10554 6adb6e08-d915-0410-941f-83917bcadc18

+++ source/ANALYSIS/OPENSWATH/MRMTransitionGroupPicker.C
@@ -264,8 +264,8 @@ namespace OpenMS
         if (picked_chroms[k][i].getIntensity() > largest)
         {
           largest = picked_chroms[k][i].getIntensity();
-          chr_idx = k;
-          peak_idx = i;
+          chr_idx = (int)k;
+          peak_idx = (int)i;
         }
       }
     }
